Coding HUD
Coding ContainerWidget, ContainerWidgetSlot, ToolTipWidget, HotbarSlotWidget, DraggedItemWidget

+++ Source/InventoryResearchDH/Public/Data/InventoryItem.h
@@ -67,7 +67,7 @@ public:
 
     // Health provided by the item
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Item")
-    int32 Health;
+    float Health;
 
     // Duration of the item
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Item")
